fix: preserve Long file counts in task progress

+++ v2/app/src/main/java/io/github/xgl34222220/baize/root/TaskCoordinator.kt
@@ -67,7 +67,7 @@ internal class TaskCoordinator(
         currentPath: String,
         startedRealtime: Long,
         deletedBytes: Long = 0L,
-        deletedFiles: Int = 0,
+        deletedFiles: Long = 0L,
         failures: Int = 0
     ) {
         setState(
